# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: HamiCloud Control Plane API
  description: >-
    Authoritative public API contract for HamiCloud distributed application runtime.
    All tenant endpoints require authentication via OIDC bearer token. Workspace access
    and permissions are derived strictly from the authenticated subject's workspace membership;
    clients never supply the workspace ID for authorization.
  version: 1.0.0
servers:
  - url: /
    description: Control Plane API
security:
  - BearerAuth: []
paths:
...
  /v1/workspaces/{workspace_id}/repositories:
    get:
      summary: List repositories in workspace
      description: Retrieve all connected Git repositories for a workspace; requires VIEWER role or higher.
      operationId: listRepositories
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: List of connected repositories
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryListResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '405':
          $ref: '#/components/responses/405MethodNotAllowed'
    post:
      summary: Connect a Git repository to workspace
      description: Connects a GitHub repository to workspace for automated builds; requires DEVELOPER role or higher.
      operationId: createRepository
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRepositoryRequest'
      responses:
        '201':
          description: Repository connected successfully
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '409':
          $ref: '#/components/responses/409Conflict'
        '422':
          $ref: '#/components/responses/422UnprocessableEntity'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '405':
          $ref: '#/components/responses/405MethodNotAllowed'
...
  /v1/repositories/{repository_id}:
    get:
      summary: Get repository details
      description: Retrieve connected repository metadata; requires VIEWER role or higher in workspace.
      operationId: getRepository
      parameters:
        - name: repository_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      responses:
        '200':
          description: Repository details
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/X-Correlation-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '405':
          $ref: '#/components/responses/405MethodNotAllowed'
  /v1/webhooks/github/{repository_id}:
    post:
      summary: GitHub Webhook receiver
      description: Ingests signed GitHub webhook push events and triggers build pipelines. Unauthenticated by bearer token; authenticated via HMAC-SHA256 signature in X-Hub-Signature-256 header.
      operationId: handleGithubWebhook
      security: []
      parameters:
        - name: repository_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema:
            type: string
            example: "72d3162e-cc78-11e3-81ab-4c9367dc0958"
        - name: X-GitHub-Event
          in: header
          required: false
          schema:
            type: string
            default: "push"
            example: "push"
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
            example: "sha256=d57c292233604f3d2f2d93e25b16952d7e10c73e04eb9df69335ef076228383a"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook processed, ping acknowledged, or duplicate delivery ignored
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  action:
                    type: string
        '202':
          description: Webhook push accepted and build requested
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  action:
                    type: string
                  release_id:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
        '405':
          $ref: '#/components/responses/405MethodNotAllowed'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OpenID Connect (OIDC) Bearer token.
  headers:
    X-Correlation-ID:
      description: Unique correlation identifier for tracing across platform components
      schema:
        type: string
        example: "req_01HVG89ABCDEF1234567890"
  responses:
    400BadRequest:
      description: Invalid request payload or query parameter
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/X-Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "BAD_REQUEST"
            message: "Invalid pagination cursor"
            correlation_id: "req_01HVG89ABCDEF1234567890"
    401Unauthorized:
      description: Authentication required or invalid credentials
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/X-Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "UNAUTHORIZED"
            message: "Authentication required"
            correlation_id: "req_01HVG89ABCDEF1234567890"
    403Forbidden:
      description: Insufficient workspace role for this operation
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/X-Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "FORBIDDEN"
            message: "Insufficient role for this operation"
            correlation_id: "req_01HVG89ABCDEF1234567890"
    404NotFound:
      description: Resource not found or not in workspace
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/X-Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            JobNotFound:
              summary: Job not found
              value:
                error_code: "NOT_FOUND"
                message: "Job not found"
                correlation_id: "req_01HVG89ABCDEF1234567890"
            WorkspaceNotFound:
              summary: Workspace not found
              value:
                error_code: "NOT_FOUND"
                message: "Workspace not found"
                correlation_id: "req_01HVG89ABCDEF1234567890"
            ApplicationNotFound:
              summary: Application not found
              value:
                error_code: "NOT_FOUND"
                message: "Application not found"
                correlation_id: "req_01HVG89ABCDEF1234567890"
            TargetReleaseNotFound:
              summary: Target release not found
              value:
                error_code: "NOT_FOUND"
                message: "Target release not found for this application"
                correlation_id: "req_01HVG89ABCDEF1234567890"
            OperationNotFound:
              summary: Operation not found
              value:
                error_code: "NOT_FOUND"
                message: "Operation not found"
                correlation_id: "req_01HVG89ABCDEF1234567890"
    405MethodNotAllowed:
      description: HTTP method not allowed for requested endpoint
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/X-Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "METHOD_NOT_ALLOWED"
            message: "Method Not Allowed"
            correlation_id: "req_01HVG89ABCDEF1234567890"
    409Conflict:
      description: State conflict or idempotency key mismatch
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/X-Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            Conflict:
              summary: Resource state or slug conflict
              value:
                error_code: "CONFLICT"
                message: "Application with slug 'payment-service' already exists in workspace"
                correlation_id: "req_01HVG89ABCDEF1234567890"
            IdempotencyConflict:
              summary: Idempotency key reused with different request payload
              value:
                error_code: "IDEMPOTENCY_CONFLICT"
                message: "Idempotency key reused with different request payload"
                correlation_id: "req_01HVG89ABCDEF1234567890"
    422UnprocessableEntity:
      description: Request payload failed schema validation
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/X-Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ValidationError:
              summary: Validation failure in request body
              value:
                error_code: "VALIDATION_ERROR"
                message: "Request validation failed"
                correlation_id: "req_01HVG89ABCDEF1234567890"
                details:
                  errors:
                    - loc: ["body", "name"]
                      msg: "Field required"
                      type: "missing"
                      input: null
            MissingIdempotencyKey:
              summary: Missing required Idempotency-Key header
              value:
                error_code: "VALIDATION_ERROR"
                message: "Request validation failed"
                correlation_id: "req_01HVG89ABCDEF1234567890"
                details:
                  errors:
                    - loc: ["header", "Idempotency-Key"]
                      msg: "Field required"
                      type: "missing"
                      input: null
    500InternalServerError:
      description: Internal platform error occurred
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/X-Correlation-ID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: "INTERNAL_SERVER_ERROR"
            message: "An unexpected internal platform error occurred."
            correlation_id: "req_01HVG89ABCDEF1234567890"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error_code, message, correlation_id]
      properties:
        error_code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - IDEMPOTENCY_CONFLICT
            - BAD_REQUEST
            - SERVICE_UNAVAILABLE
            - NOT_IMPLEMENTED
            - INTERNAL_SERVER_ERROR
            - IMAGE_POLICY_VIOLATION
          example: "NOT_FOUND"
        message:
          type: string
          example: "Job not found"
        correlation_id:
          type: string
          example: "req_01HVG89ABCDEF1234567890"
        details:
          type: [object, 'null']
          additionalProperties: true
      example:
        error_code: "NOT_FOUND"
        message: "Job not found"
        correlation_id: "req_01HVG89ABCDEF1234567890"
...
    CreateRepositoryRequest:
      type: object
      required: [name, repo_url, webhook_secret]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "payment-service"
        repo_url:
          type: string
          minLength: 1
          maxLength: 255
          example: "https://github.com/example-org/payment-service"
        webhook_secret:
          type: string
          minLength: 1
          maxLength: 255
          example: "whsec_01HVG89ABCDEF1234567890"
        default_branch:
          type: string
          maxLength: 100
          default: "main"
          example: "main"
      example:
        name: "payment-service"
        repo_url: "https://github.com/example-org/payment-service"
        webhook_secret: "whsec_01HVG89ABCDEF1234567890"
        default_branch: "main"
    RepositoryResponse:
      type: object
      required: [id, workspace_id, name, repo_url, default_branch, created_at]
      properties:
        id:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        workspace_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "payment-service"
        repo_url:
          type: string
          example: "https://github.com/example-org/payment-service"
        default_branch:
          type: string
          example: "main"
        created_at:
          type: string
          format: date-time
          example: "2026-09-19T12:00:00Z"
      example:
        id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        workspace_id: "550e8400-e29b-41d4-a716-446655440000"
        name: "payment-service"
        repo_url: "https://github.com/example-org/payment-service"
        default_branch: "main"
        created_at: "2026-09-19T12:00:00Z"
    RepositoryListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RepositoryResponse'
        next_cursor:
          type: [string, 'null']
          example: null
      example:
        items:
          - id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
            workspace_id: "550e8400-e29b-41d4-a716-446655440000"
            name: "payment-service"
            repo_url: "https://github.com/example-org/payment-service"
            default_branch: "main"
            created_at: "2026-09-19T12:00:00Z"
        next_cursor: null
...